refactor to DDD-esq folder structure
