+++ Makefile
@@ -4,7 +4,7 @@ DOCKER_COMPOSE = IMAGE_TAG=${IMAGE_TAG} docker-compose -f docker-compose.build.y
 DOCKER_COMPOSE_TEST = docker-compose -f docker-compose.test.yml
 DOCKER_COMPOSE_XPUB_POSTGRES = IMAGE_TAG=${IMAGE_TAG} docker-compose -f docker-compose.xpub-postgres.yml
 PUSH_COMMAND = IMAGE_TAG=${IMAGE_TAG} .scripts/travis/push-image.sh
-GET_SCHEMA_TABLES = psql -q -t -U postgres -c "select count(*) from information_schema.tables where table_schema='xpublegacy';" | xargs
+GET_SCHEMA_TABLES = psql -q -t -U postgres postgres -c "select count(*) from information_schema.tables where table_schema='public';" | xargs
 LOAD_SCHEMA = psql -U postgres -f xpub-schema.sql
 
 setup:
@@ -27,7 +27,7 @@ test_integration:
 	./.scripts/docker/wait-healthy.sh test_postgres 20
 	${DOCKER_COMPOSE_TEST} up -d application
 	./.scripts/docker/wait-healthy.sh test_reviewer-submission 20
-	yarn run test:integration
+	CONFIG_PATH=./config/config.json yarn run test:integration
 	${DOCKER_COMPOSE_TEST} down
 
 load_schema:
