--- src/resolvers/index.ts
@@ -1,3 +1,0 @@
-export * from './submission';
-export * from './survey';
-export * from './user';
