+++ src/submission/resolvers/submission.ts
@@ -1,5 +1,5 @@
-import { SubmissionService } from '../services/submission';
-import { SubmissionId, DtoViewSubmission } from '../types/submission';
+import { SubmissionService } from '../models/submission-service';
+import { SubmissionId, DtoViewSubmission } from '../submission';
 import { IResolvers } from 'apollo-server-express';
 
 const resolvers = (submissionService: SubmissionService): IResolvers => ({
