+++ src/survey/models/answer.test.ts
@@ -1,0 +1,3 @@
+describe('Answer Entity', () => {
+    it('answer is good', async () => {});
+});
