+++ src/survey/models/survey-response.test.ts
@@ -1,7 +1,7 @@
 import { v4 as uuid } from 'uuid';
 import { SurveyResponse } from './survey-response';
-import { SurveyId, SurveyResponseId } from '../types/survey';
-import { SubmissionId } from '../types/submission';
+import { SurveyId, SurveyResponseId } from '../survey';
+import { SubmissionId } from '../../submission/submission';
 import { Question } from './question';
 import { Answer } from './answer';
