+++ src/user/infrastructure/rest-user.ts
@@ -1,5 +1,5 @@
-import { InfraLogger as logger } from '../logger';
-import { User, UserRepository } from '../types/user';
+import { InfraLogger as logger } from '../../logger';
+import { User, UserRepository } from '../user';
 import fetch from 'node-fetch';
 
 export class RestUserRepository implements UserRepository {
